Update wiki
Update project folder structure

Move Getting-started into the project README (which already duplicated
most of the information). There was some information there about
installing via yum and rpm. If that is useful, perhaps it belongs on a
page about deploying to production?

Move content of 'Deployment process' and 'Implementation decisions' to
new 'Components' page, since that is what the page is really about. Much
of the components page still needs to be fleshed out, but I tried to
create stubs for most things.

Light editing of 'Background' and 'Explaining the Stack', including
updating to reflect release of ML 8.

REST extensions page still needs to be filled out

Add an initial 'Customization recipes' page with a recipe on adding a
facet.